+++ java/src/test/java/io/cucumber/query/QueryAcceptanceTest.java
@@ -195,7 +195,13 @@ public class QueryAcceptanceTest {
                 .map(Convertor::toMessage));
         results.put("findTestRunFinished", query.findTestRunFinished());
         results.put("findTestRunStarted", query.findTestRunStarted());
-        results.put("findTestStepBy", query.findAllTestCaseStarted().stream()
+        results.put("findTestStepByTestStepStarted", query.findAllTestCaseStarted().stream()
+                .map(query::findTestStepsStartedBy)
+                .flatMap(Collection::stream)
+                .map(query::findTestStepBy)
+                .map(testStep -> testStep.map(TestStep::getId))
+                .collect(toList()));
+        results.put("findTestStepByTestStepFinished", query.findAllTestCaseStarted().stream()
                 .map(query::findTestStepsFinishedBy)
                 .flatMap(Collection::stream)
                 .map(query::findTestStepBy)