+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- [Java] Make `Lineage` API public and add `Query.findLineageBy` methods 
 
 ## [13.2.0] - 2025-02-02
 ### Changed